+++ Drone.cpp
@@ -1,29 +1,3 @@
-/*
-
-This library is free software; you can redistribute it and/or
-modify it under the terms of the GNU Lesser General Public
-License as published by the Free Software Foundation; either
-version 2.1 of the License, or (at your option) any later version.
-
-This library is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public
-License along with this library; if not, write to the Free Software
-Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
-
-The latest version of this library can always be found at
-http://arduiniana.org.
-*/
-
-// When set, _DEBUG co-opts pins 11 and 13 for debugging with an
-// oscilloscope or logic analyzer.  Beware: it also slightly modifies
-// the bit times, so don't rely on it too much at high baud rates
-#define _DEBUG 0
-#define _DEBUG_PIN1 11
-#define _DEBUG_PIN2 13
 // 
 // Includes
 // 
@@ -97,299 +71,6 @@ void Drone::arm() {
 
 
 
-// ***********************************************
-// SerialIO Class Implementation
-// ***********************************************
-SerialIO::SerialIO() {}
-
-SerialIO::SerialIO(int txPin , int rxPin) {
-  softwareSerial = SoftwareSerial(txPin, rxPin);
-  softwareSerial.begin(57600);
-  selectedSerialPort = swSerialPort;
-}
-
-SerialIO::SerialIO(SerialPort serialPort) {
-  selectedSerialPort = serialPort;
-  switch(serialPort) {
-    case serialPort0:
-      #if defined(UBRRH) || defined(UBRR0H) || defined(UBRR1H)
-        Serial.begin(57600);
-      #endif
-      break;
-    case serialPort1:
-      #if defined(UBRR1H)
-        Serial1.begin(57600);
-      #endif
-      break;
-    case serialPort2:
-      #if defined(UBRR2H)
-        Serial2.begin(57600);
-      #endif
-      break;
-    case serialPort3:
-      #if defined(UBRR3H)
-        Serial3.begin(57600);
-      #endif
-      break;
-    default:
-      Serial.println("Error: bad argument to SerialIO constructor");
-      break;
-  }
-}
-
-void SerialIO::write(uint8_t data) {
-  switch(selectedSerialPort) {
-    case serialPort0:
-      #if defined(UBRRH) || defined(UBRR0H) || defined(UBRR1H)
-        Serial.write(data);
-      #endif
-      break;
-    case serialPort1:
-      #if defined(UBRR1H)
-        Serial1.write(data);
-      #endif
-      break;
-    case serialPort2:
-      #if defined(UBRR2H)
-        Serial2.write(data);
-      #endif
-      break;
-    case serialPort3:
-      #if defined(UBRR3H)
-        Serial3.write(data);
-      #endif
-      break;
-    case swSerialPort:
-      softwareSerial.write(data);
-      break;
-    default:
-      Serial.println("Error: incorrectly configured serial settings.");
-  }
-}
-
-uint8_t SerialIO::serialRead() {
-  switch(selectedSerialPort) {
-    case serialPort0:
-      #if defined(UBRRH) || defined(UBRR0H) || defined(UBRR1H)
-        return Serial.read();
-      #endif
-      break;
-    case serialPort1:
-      #if defined(UBRR1H)
-        return Serial1.read();
-      #endif
-      break;
-    case serialPort2:
-      #if defined(UBRR2H)
-        return Serial2.read();
-      #endif
-      break;
-    case serialPort3:
-      #if defined(UBRR3H)
-        return Serial3.read();
-      #endif
-      break;
-    case swSerialPort:
-      return softwareSerial.read();
-      break;
-    default:
-      Serial.println("Error: incorrectly configured serial settings.");
-  }
-}
-
-bool SerialIO::serialAvailable() {
-  switch(selectedSerialPort) {
-    case serialPort0:
-      #if defined(UBRRH) || defined(UBRR0H) || defined(UBRR1H)
-        return Serial.available();
-      #endif
-      break;
-    case serialPort1:
-      #if defined(UBRR1H)
-        return Serial1.available();
-      #endif
-      break;
-    case serialPort2:
-      #if defined(UBRR2H)
-        return Serial2.available();
-      #endif
-      break;
-    case serialPort3:
-      #if defined(UBRR3H)
-        return Serial3.available();
-      #endif
-      break;
-    case swSerialPort:
-      return softwareSerial.available();
-      break;
-    default:
-      Serial.println("Error: incorrectly configured serial settings.");
-  }
-}
-
-void SerialIO::sendPacket(uint8_t data, uint8_t resourceID, uint8_t actionID) {
-  write('$');
-  write(1);
-  write(resourceID);
-  write(actionID);
-  write(data);
-  write((1+resourceID+actionID+data)%256);
-}
-
-void SerialIO::sendPacket(int8_t data, uint8_t resourceID, uint8_t actionID) {
-  write('$');
-  write(1);
-  write(byte(resourceID));
-  write(byte(actionID));
-  write(data);
-  // Cast data because the firmware interprets packets as unsigned for calculating checksum.
-  write((1+resourceID+actionID+(uint8_t)data)%256);
-}
-
-void SerialIO::sendPacket(int16_t data, uint8_t resourceID, uint8_t actionID) {
-  write('$');
-  write(2);
-  write(byte(resourceID));
-  write(byte(actionID));
-  write(highByte(data));
-  write(lowByte(data));
-  write((2+resourceID+actionID+highByte(data)+lowByte(data))%256);
-}
-
-void SerialIO::sendPacket(float data, uint8_t resourceID, uint8_t actionID) {
-  union u_tag {
-    uint8_t b[4];
-    float dataFloat;
-  } u;
-  u.dataFloat = data;
-
-  write('$');
-  write(4);
-  write(byte(resourceID));
-  write(byte(actionID));
-  write(u.b[0]);
-  write(u.b[1]);
-  write(u.b[2]);
-  write(u.b[3]);
-  write((4+resourceID+actionID+u.b[0]+u.b[1]+u.b[2]+u.b[3])%256);
-}
-
-void SerialIO::print(String data) {
-  switch(selectedSerialPort) {
-    case serialPort0:
-      #if defined(UBRRH) || defined(UBRR0H) || defined(UBRR1H)
-        Serial.print(data);
-      #endif
-      break;
-    case serialPort1:
-      #if defined(UBRR1H)
-        Serial1.print(data);
-      #endif
-      break;
-    case serialPort2:
-      #if defined(UBRR2H)
-        Serial2.print(data);
-      #endif
-      break;
-    case serialPort3:
-      #if defined(UBRR3H)
-        Serial3.print(data);
-      #endif
-      break;
-    case swSerialPort:
-      softwareSerial.print(data);
-      break;
-    default:
-      Serial.println("Error: incorrectly configured serial settings.");
-  }
-}
-
-void SerialIO::readBytes(char *buffer, int bytesToRead) {
-/*
-  if (isHwSerial0Used) {
-    #if defined(UBRRH) || defined(UBRR0H) || defined(UBRR1H)
-      Serial.readBytes(buffer, bytesToRead);
-    #endif
-  } else if (isHwSerial1Used) {
-    #if defined(UBRR1H)
-      Serial1.readBytes(buffer, bytesToRead);
-    #endif
-  } else if (isHwSerial2Used) {
-    #if defined(UBRR2H)
-      Serial2.readBytes(buffer, bytesToRead);
-    #endif
-  } else if (isHwSerial3Used) {
-    #if defined(UBRR3H)
-        Serial3.readBytes(buffer, bytesToRead);
-    #endif
-  } else if (isSwSerialUsed) {
-    softwareSerial.readBytes(buffer, bytesToRead);
-  }
-*/
-}
-// ***********************************************
-// AvantRC Class Implementation
-// ***********************************************
-AvantRC::AvantRC() {};
-AvantRC::AvantRC(SerialIO *rcTservice, Callback *callback) {
-  service = rcTservice;
-  myCallback = callback;
-}
-
-void AvantRC::setAileron(int8_t value){
-  service->sendPacket(value, 2, 4);
-};
-void AvantRC::setElevator(int8_t value){
-  service->sendPacket(value, 2, 3);
-};
-void AvantRC::setThrottle(int8_t value){
-  service->sendPacket(value, 2, 2);
-};
-void AvantRC::setRudder(int8_t value){
-  service->sendPacket(value, 2, 1);
-}; 
-void AvantRC::setFlightMode(int8_t value){
-  service->sendPacket(value, 2, 5);
-};
-void AvantRC::getAileron(){
-  service->sendPacket((int16_t)0, 2, 9);
-}
-void AvantRC::getElevator(){
-  service->sendPacket((int16_t)0, 2, 8);
-}
-void AvantRC::getThrottle(){
-  service->sendPacket((int16_t)0, 2, 7);
-}
-void AvantRC::getRudder(){
-  service->sendPacket((int16_t)0, 2, 6);
-}
-void AvantRC::getFlightMode(){
-  service->sendPacket((int16_t)0, 2, 10);
-}
-
-void AvantRC::flightModeCallback(void (*function)(byte)) {
-  (*myCallback).flightMode = function;
-}
-
-void AvantRC::elevatorCallback(void (*function)(byte)) {
-  (*myCallback).elevator= function;
-}
-
-void AvantRC::aileronCallback(void (*function)(byte)) {
-  (*myCallback).aileron = function;
-}
-
-void AvantRC::rudderCallback(void (*function)(byte)) {
-  (*myCallback).rudder = function;
-}
-
-void AvantRC::throttleCallback(void (*function)(byte)) {
-  (*myCallback).throttle = function;
-}
-
-void AvantRC::sendRTEA(uint8_t rudder, uint8_t throttle, uint8_t elevator, uint8_t aileron){
-  long data = (long(rudder+100))+ (long(throttle+100) << 8) + (long(elevator+100) << 16) + (long(aileron+100) << 24);
-}
 
 
 //************************************************
@@ -989,446 +670,4 @@ void AvantResponseHandler::responseHandler() {
       Serial.println("checksum failed");
     }//end of if Serial available
   }//end of Serial.read
-}
-
-//*****************************************************
-//SoftwareSerial Code
-//
-// Lookup table
-//
-//\cond
-typedef struct _DELAY_TABLE
-//\endcond
-{
-  long baud;
-  unsigned short rx_delay_centering;
-  unsigned short rx_delay_intrabit;
-  unsigned short rx_delay_stopbit;
-  unsigned short tx_delay;
-} DELAY_TABLE;
-
-#if F_CPU == 16000000
-
-static const DELAY_TABLE PROGMEM table[] = 
-{
-  //  baud    rxcenter   rxintra    rxstop    tx
-  { 115200,   1,         17,        17,       12,    },
-  { 57600,    10,        37,        37,       33,    },
-  { 38400,    25,        57,        57,       54,    },
-  { 31250,    31,        70,        70,       68,    },
-  { 28800,    34,        77,        77,       74,    },
-  { 19200,    54,        117,       117,      114,   },
-  { 14400,    74,        156,       156,      153,   },
-  { 9600,     114,       236,       236,      233,   },
-  { 4800,     233,       474,       474,      471,   },
-  { 2400,     471,       950,       950,      947,   },
-  { 1200,     947,       1902,      1902,     1899,  },
-  { 600,      1902,      3804,      3804,     3800,  },
-  { 300,      3804,      7617,      7617,     7614,  },
-};
-
-const int XMIT_START_ADJUSTMENT = 5;
-
-#elif F_CPU == 8000000
-
-static const DELAY_TABLE table[] PROGMEM = 
-{
-  //  baud    rxcenter    rxintra    rxstop  tx
-  { 115200,   1,          5,         5,      3,      },
-  { 57600,    1,          15,        15,     13,     },
-  { 38400,    2,          25,        26,     23,     },
-  { 31250,    7,          32,        33,     29,     },
-  { 28800,    11,         35,        35,     32,     },
-  { 19200,    20,         55,        55,     52,     },
-  { 14400,    30,         75,        75,     72,     },
-  { 9600,     50,         114,       114,    112,    },
-  { 4800,     110,        233,       233,    230,    },
-  { 2400,     229,        472,       472,    469,    },
-  { 1200,     467,        948,       948,    945,    },
-  { 600,      948,        1895,      1895,   1890,   },
-  { 300,      1895,       3805,      3805,   3802,   },
-};
-
-const int XMIT_START_ADJUSTMENT = 4;
-
-#elif F_CPU == 20000000
-
-// 20MHz support courtesy of the good people at macegr.com.
-// Thanks, Garrett!
-
-static const DELAY_TABLE PROGMEM table[] =
-{
-  //  baud    rxcenter    rxintra    rxstop  tx
-  { 115200,   3,          21,        21,     18,     },
-  { 57600,    20,         43,        43,     41,     },
-  { 38400,    37,         73,        73,     70,     },
-  { 31250,    45,         89,        89,     88,     },
-  { 28800,    46,         98,        98,     95,     },
-  { 19200,    71,         148,       148,    145,    },
-  { 14400,    96,         197,       197,    194,    },
-  { 9600,     146,        297,       297,    294,    },
-  { 4800,     296,        595,       595,    592,    },
-  { 2400,     592,        1189,      1189,   1186,   },
-  { 1200,     1187,       2379,      2379,   2376,   },
-  { 600,      2379,       4759,      4759,   4755,   },
-  { 300,      4759,       9523,      9523,   9520,   },
-};
-
-const int XMIT_START_ADJUSTMENT = 6;
-
-#else
-
-#error This version of SoftwareSerial supports only 20, 16 and 8MHz processors
-
-#endif
-
-//
-// Statics
-//
-SoftwareSerial *SoftwareSerial::active_object = 0;
-char SoftwareSerial::_receive_buffer[_SS_MAX_RX_BUFF]; 
-volatile uint8_t SoftwareSerial::_receive_buffer_tail = 0;
-volatile uint8_t SoftwareSerial::_receive_buffer_head = 0;
-
-//
-// Debugging
-//
-// This function generates a brief pulse
-// for debugging or measuring on an oscilloscope.
-inline void DebugPulse(uint8_t pin, uint8_t count) {
-#if _DEBUG
-  volatile uint8_t *pport = portOutputRegister(digitalPinToPort(pin));
-
-  uint8_t val = *pport;
-  while (count--) {
-    *pport = val | digitalPinToBitMask(pin);
-    *pport = val;
-  }
-#endif
-}
-
-//
-// Private methods
-//
-
-/* static */ 
-inline void SoftwareSerial::tunedDelay(uint16_t delay) { 
-  uint8_t tmp=0;
-
-  asm volatile("sbiw    %0, 0x01 \n\t"
-    "ldi %1, 0xFF \n\t"
-    "cpi %A0, 0xFF \n\t"
-    "cpc %B0, %1 \n\t"
-    "brne .-10 \n\t"
-    : "+r" (delay), "+a" (tmp)
-    : "0" (delay)
-    );
-}
-
-// This function sets the current object as the "listening"
-// one and returns true if it replaces another 
-bool SoftwareSerial::listen() {
-  if (active_object != this) {
-    _buffer_overflow = false;
-    uint8_t oldSREG = SREG;
-    cli();
-    _receive_buffer_head = _receive_buffer_tail = 0;
-    active_object = this;
-    SREG = oldSREG;
-    return true;
-  }
-
-  return false;
-}
-
-//
-// The receive routine called by the interrupt handler
-//
-void SoftwareSerial::recv() {
-
-#if GCC_VERSION < 40302
-// Work-around for avr-gcc 4.3.0 OSX version bug
-// Preserve the registers that the compiler misses
-// (courtesy of Arduino forum user *etracer*)
-  asm volatile(
-    "push r18 \n\t"
-    "push r19 \n\t"
-    "push r20 \n\t"
-    "push r21 \n\t"
-    "push r22 \n\t"
-    "push r23 \n\t"
-    "push r26 \n\t"
-    "push r27 \n\t"
-    ::);
-#endif  
-
-  uint8_t d = 0;
-
-  // If RX line is high, then we don't see any start bit
-  // so interrupt is probably not for us
-  if (_inverse_logic ? rx_pin_read() : !rx_pin_read()) {
-    // Wait approximately 1/2 of a bit width to "center" the sample
-    tunedDelay(_rx_delay_centering);
-    DebugPulse(_DEBUG_PIN2, 1);
-
-    // Read each of the 8 bits
-    for (uint8_t i=0x1; i; i <<= 1) {
-      tunedDelay(_rx_delay_intrabit);
-      DebugPulse(_DEBUG_PIN2, 1);
-      uint8_t noti = ~i;
-      if (rx_pin_read())
-        d |= i;
-      else // else clause added to ensure function timing is ~balanced
-        d &= noti;
-    }
-
-    // skip the stop bit
-    tunedDelay(_rx_delay_stopbit);
-    DebugPulse(_DEBUG_PIN2, 1);
-
-    if (_inverse_logic)
-      d = ~d;
-
-    // if buffer full, set the overflow flag and return
-    if ((_receive_buffer_tail + 1) % _SS_MAX_RX_BUFF != _receive_buffer_head) {
-      // save new data in buffer: tail points to where byte goes
-      _receive_buffer[_receive_buffer_tail] = d; // save new byte
-      _receive_buffer_tail = (_receive_buffer_tail + 1) % _SS_MAX_RX_BUFF;
-    } else {
-    #if _DEBUG // for scope: pulse pin as overflow indictator
-      DebugPulse(_DEBUG_PIN1, 1);
-    #endif
-      _buffer_overflow = true;
-    }
-  }
-
-#if GCC_VERSION < 40302
-// Work-around for avr-gcc 4.3.0 OSX version bug
-// Restore the registers that the compiler misses
-  asm volatile(
-    "pop r27 \n\t"
-    "pop r26 \n\t"
-    "pop r23 \n\t"
-    "pop r22 \n\t"
-    "pop r21 \n\t"
-    "pop r20 \n\t"
-    "pop r19 \n\t"
-    "pop r18 \n\t"
-    ::);
-#endif
-}
-
-void SoftwareSerial::tx_pin_write(uint8_t pin_state) {
-  if (pin_state == LOW)
-    *_transmitPortRegister &= ~_transmitBitMask;
-  else
-    *_transmitPortRegister |= _transmitBitMask;
-}
-
-uint8_t SoftwareSerial::rx_pin_read() {
-  return *_receivePortRegister & _receiveBitMask;
-}
-
-//
-// Interrupt handling
-//
-
-/* static */
-inline void SoftwareSerial::handle_interrupt() {
-  if (active_object) {
-    active_object->recv();
-  }
-}
-
-#if defined(PCINT0_vect)
-ISR(PCINT0_vect) {
-  SoftwareSerial::handle_interrupt();
-}
-#endif
-
-#if defined(PCINT1_vect)
-ISR(PCINT1_vect) {
-  SoftwareSerial::handle_interrupt();
-}
-#endif
-
-#if defined(PCINT2_vect)
-ISR(PCINT2_vect) {
-  SoftwareSerial::handle_interrupt();
-}
-#endif
-
-#if defined(PCINT3_vect)
-ISR(PCINT3_vect) {
-  SoftwareSerial::handle_interrupt();
-}
-#endif
-
-//
-// Constructor
-//
-SoftwareSerial::SoftwareSerial(uint8_t receivePin, uint8_t transmitPin, bool inverse_logic /* = false */) : 
-  _rx_delay_centering(0),
-  _rx_delay_intrabit(0),
-  _rx_delay_stopbit(0),
-  _tx_delay(0),
-  _buffer_overflow(false),
-  _inverse_logic(inverse_logic)
-{
-  setTX(transmitPin);
-  setRX(receivePin);
-}
-
-//
-// Destructor
-//
-SoftwareSerial::~SoftwareSerial() {
-  end();
-}
-
-void SoftwareSerial::setTX(uint8_t tx) {
-  pinMode(tx, OUTPUT);
-  digitalWrite(tx, HIGH);
-  _transmitBitMask = digitalPinToBitMask(tx);
-  uint8_t port = digitalPinToPort(tx);
-  _transmitPortRegister = portOutputRegister(port);
-}
-
-void SoftwareSerial::setRX(uint8_t rx) {
-  pinMode(rx, INPUT);
-  if (!_inverse_logic)
-    digitalWrite(rx, HIGH);  // pullup for normal logic!
-  _receivePin = rx;
-  _receiveBitMask = digitalPinToBitMask(rx);
-  uint8_t port = digitalPinToPort(rx);
-  _receivePortRegister = portInputRegister(port);
-}
-
-//
-// Public methods
-//
-
-void SoftwareSerial::begin(long speed) {
-  _rx_delay_centering = _rx_delay_intrabit = _rx_delay_stopbit = _tx_delay = 0;
-
-  for (unsigned i=0; i<sizeof(table)/sizeof(table[0]); ++i) {
-    long baud = pgm_read_dword(&table[i].baud);
-    if (baud == speed) {
-      _rx_delay_centering = pgm_read_word(&table[i].rx_delay_centering);
-      _rx_delay_intrabit = pgm_read_word(&table[i].rx_delay_intrabit);
-      _rx_delay_stopbit = pgm_read_word(&table[i].rx_delay_stopbit);
-      _tx_delay = pgm_read_word(&table[i].tx_delay);
-      break;
-    }
-  }
-
-  // Set up RX interrupts, but only if we have a valid RX baud rate
-  if (_rx_delay_stopbit) {
-    if (digitalPinToPCICR(_receivePin)) {
-      *digitalPinToPCICR(_receivePin) |= _BV(digitalPinToPCICRbit(_receivePin));
-      *digitalPinToPCMSK(_receivePin) |= _BV(digitalPinToPCMSKbit(_receivePin));
-    }
-    tunedDelay(_tx_delay); // if we were low this establishes the end
-  }
-
-#if _DEBUG
-  pinMode(_DEBUG_PIN1, OUTPUT);
-  pinMode(_DEBUG_PIN2, OUTPUT);
-#endif
-
-  listen();
-}
-
-void SoftwareSerial::end() {
-  if (digitalPinToPCMSK(_receivePin))
-    *digitalPinToPCMSK(_receivePin) &= ~_BV(digitalPinToPCMSKbit(_receivePin));
-}
-
-
-// Read data from buffer
-int SoftwareSerial::read() {
-  if (!isListening())
-    return -1;
-
-  // Empty buffer?
-  if (_receive_buffer_head == _receive_buffer_tail)
-    return -1;
-
-  // Read from "head"
-  uint8_t d = _receive_buffer[_receive_buffer_head]; // grab next byte
-  _receive_buffer_head = (_receive_buffer_head + 1) % _SS_MAX_RX_BUFF;
-  return d;
-}
-
-int SoftwareSerial::available() {
-  if (!isListening())
-    return 0;
-
-  return (_receive_buffer_tail + _SS_MAX_RX_BUFF - _receive_buffer_head) % _SS_MAX_RX_BUFF;
-}
-
-size_t SoftwareSerial::write(uint8_t b) {
-  if (_tx_delay == 0) {
-    setWriteError();
-    return 0;
-  }
-
-  uint8_t oldSREG = SREG;
-  cli();  // turn off interrupts for a clean txmit
-
-  // Write the start bit
-  tx_pin_write(_inverse_logic ? HIGH : LOW);
-  tunedDelay(_tx_delay + XMIT_START_ADJUSTMENT);
-
-  // Write each of the 8 bits
-  if (_inverse_logic) {
-    for (byte mask = 0x01; mask; mask <<= 1) {
-      if (b & mask) // choose bit
-        tx_pin_write(LOW); // send 1
-      else
-        tx_pin_write(HIGH); // send 0
-    
-      tunedDelay(_tx_delay);
-    }
-
-    tx_pin_write(LOW); // restore pin to natural state
-  } else {
-    for (byte mask = 0x01; mask; mask <<= 1) {
-      if (b & mask) // choose bit
-        tx_pin_write(HIGH); // send 1
-      else
-        tx_pin_write(LOW); // send 0
-    
-      tunedDelay(_tx_delay);
-    }
-
-    tx_pin_write(HIGH); // restore pin to natural state
-  }
-
-  SREG = oldSREG; // turn interrupts back on
-  tunedDelay(_tx_delay);
-  
-  return 1;
-}
-
-void SoftwareSerial::flush() {
-  if (!isListening())
-    return;
-
-  uint8_t oldSREG = SREG;
-  cli();
-  _receive_buffer_head = _receive_buffer_tail = 0;
-  SREG = oldSREG;
-}
-
-int SoftwareSerial::peek() {
-  if (!isListening())
-    return -1;
-
-  // Empty buffer?
-  if (_receive_buffer_head == _receive_buffer_tail)
-    return -1;
-
-  // Read from "head"
-  return _receive_buffer[_receive_buffer_head];
 }
