+++ Drone.h
@@ -21,235 +21,22 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include <inttypes.h>
 #include <Stream.h>
 #include <Arduino.h>
+#include "AvantRC.h"
+#include "SerialIO.h"
+#include "Callback.h"
 
 /******************************************************************************
 * Definitions
 ******************************************************************************/
 
-#define _SS_MAX_RX_BUFF 256 // RX buffer size
-#ifndef GCC_VERSION
-#define GCC_VERSION (__GNUC__ * 10000 + __GNUC_MINOR__ * 100 + __GNUC_PATCHLEVEL__)
-#endif
-
-// TODO: Description
-enum SerialPort {
-  serialPort0,
-  serialPort1,
-  serialPort2,
-  serialPort3,
-  swSerialPort
-};
-
-//********************************************
-//SoftwareSerial Code
-//*****************************************
-//\cond
-class SoftwareSerial : public Stream
-//\endcond
-{
-  private:
-    // per object data
-    uint8_t _receivePin;
-    uint8_t _receiveBitMask;
-    volatile uint8_t *_receivePortRegister;
-    uint8_t _transmitBitMask;
-    volatile uint8_t *_transmitPortRegister;
-
-    uint16_t _rx_delay_centering;
-    uint16_t _rx_delay_intrabit;
-    uint16_t _rx_delay_stopbit;
-    uint16_t _tx_delay;
-
-    uint16_t _buffer_overflow:1;
-    uint16_t _inverse_logic:1;
-
-    // static data
-    static char _receive_buffer[_SS_MAX_RX_BUFF]; 
-    static volatile uint8_t _receive_buffer_tail;
-    static volatile uint8_t _receive_buffer_head;
-    static SoftwareSerial *active_object;
-
-    // private methods
-    void recv();
-    uint8_t rx_pin_read();
-    void tx_pin_write(uint8_t pin_state);
-    void setTX(uint8_t transmitPin);
-    void setRX(uint8_t receivePin);
-
-    // private static method for timing
-    static inline void tunedDelay(uint16_t delay);
-
-  public:
-    // public methods
-    SoftwareSerial(){};
-    SoftwareSerial(uint8_t receivePin, uint8_t transmitPin, bool inverse_logic = false);
-    ~SoftwareSerial();
-    void begin(long speed);
-    bool listen();
-    void end();
-    bool isListening() { return this == active_object; }
-    bool overflow() { bool ret = _buffer_overflow; _buffer_overflow = false; return ret; }
-    int peek();
-
-    virtual size_t write(uint8_t byte);
-    virtual int read();
-    virtual int available();
-    virtual void flush();
-
-    using Print::write;
-
-    // public only for easy access by interrupt handlers
-    static inline void handle_interrupt();
-};
-
-// Arduino 0012 workaround
-#undef int
-#undef char
-#undef long
-#undef byte
-#undef float
-#undef abs
-#undef round
 
 
 //*****************************************
 //Avant Classes
 //*****************************************
 //\cond
-class Callback 
-//\endcond
-{
-  public:
-    //SPI Callbacks
-    void (*transfer)(byte);
-    //I2C Callbacks
-    void (*i2cRead)(byte); 
-    //GPS callbacks
-    void (*longitude)(float);
-    void (*latitude)(float);
-    void (*altitude)(float);
-    void (*speed)(float);
-    void (*satellite)(byte);
-    void (*orientation)(float);
-    //RC Callbacks
-    void (*flightMode)(byte);
-    void (*elevator)(byte);
-    void (*aileron)(byte);
-    void (*throttle)(byte);
-    void (*rudder)(byte);
-    //pulseIn Callbacks
-    void (*pulseIn1)(long);
-    void (*pulseIn2)(long);
-    void (*pulseIn3)(long);
-    void (*pulseIn4)(long);
-    void (*pulseIn5)(long);
-    void (*pulseIn6)(long);
-    void (*pulseIn7)(long);
-    void (*pulseIn8)(long);
-    void (*pulseIn9)(long);
-    void (*pulseIn10)(long);
-    //digital Read Callbacks
-    void (*digitalRead1)(byte);
-    void (*digitalRead2)(byte);
-    void (*digitalRead3)(byte);
-    void (*digitalRead4)(byte);
-    void (*digitalRead5)(byte);
-    void (*digitalRead6)(byte);
-    void (*digitalRead7)(byte);
-    void (*digitalRead8)(byte);
-    void (*digitalRead9)(byte);
-    void (*digitalRead10)(byte);
-    //analog Read Callbacks
-    void (*analogRead1)(byte);
-    void (*analogRead2)(byte);
-    void (*analogRead3)(byte);
-    void (*analogRead4)(byte);
-    //AutoPilot Callbacks
-    void (*getWaypointLatitude)(float);
-    void (*getWaypointLongitude)(float);
-    void (*getWaypointAltitude)(float);
-    void (*getWaypointOrientation)(float);
-    void (*getElevatorKP)(float);
-    void (*getElevatorKD)(float);
-    void (*getElevatorKI)(float);
-    void (*getThrottleKP)(float);
-    void (*getThrottleKD)(float);
-    void (*getThrottleKI)(float);
-    void (*getAileronKP)(float);
-    void (*getAileronKD)(float);
-    void (*getAileronKI)(float);
-    void (*getYawKP)(float);
-    void (*getYawKD)(float);
-    void (*getYawKI)(float);
-    
-};
+
   
-//\cond
-class SerialIO
-//\endcond
-{
-  friend class AvantResponseHandler;
-  private:
-    SerialPort selectedSerialPort;
-
-
-  public:
-    //TODO: change it to initialize function 
-    SoftwareSerial softwareSerial;
-    SerialIO();
-
-    /**
-    TODO
-
-    @param txPin
-    @param rxPin
-    */
-    SerialIO(int txPin , int rxPin);
-
-    /**
-    TODO
-
-    @param hwSerialCode
-    */
-    SerialIO(SerialPort serialPort);
-
-    /**
-    TODO
-
-    @param data
-    */
-    void write(uint8_t data);
-    
-    /**
-    Method description
-    */
-    uint8_t serialRead();
-    
-    /**
-    Method description
-    */
-    bool serialAvailable();
-
-
-    void sendPacket(int16_t data, uint8_t resourceID, uint8_t actionID);
-
-    void sendPacket(uint8_t data, uint8_t resourceID, uint8_t actionID);
-
-    void sendPacket(int8_t data, uint8_t resourceID, uint8_t actionID);
-
-    void sendPacket(float data, uint8_t resourceID, uint8_t actionID);
-
-    //TODO: REMOVE // DEBUG
-    /**
-    Method description
-    */
-    void print(String data);
-
-    /**
-    Method description
-    */
-    void readBytes(char *buffer, int bytesToRead);
-};
 
 //\cond
 class AvantResponseHandler
@@ -388,149 +175,6 @@ class AvantPose
          
     */
     void orientationCallback(void (*function)(float));
-};
-
-class AvantRC //handles sending values to the PWM/PPM port(s) 
-/**
- 
- This class allows you to control your drone: set the throttle, rudder, elevator, and aileron channels on the drone.
- Essentially controlling up down left right ect.
- 
-*/
-{
-  private:
-    //\cond
-    SerialIO *service;
-    Callback *myCallback;
-    //\endcond
-  public:
-    //\cond
-    AvantRC();
-    
-    AvantRC(SerialIO *rcTservice, Callback *callback);
-    //\endcond
-    
-    /**
-    
-    Sends a request to the extender to set the Aileron channel to a specific channel between -100 and 100.  
-    
-    @param value The value to set the Aileron channel to
-
-    */
-    void setAileron(int8_t value);
-    
-    /**
-         
-    Sends a request to the extender to set the Extender channel to a specific channel between -100 and 100.  
-    
-    @param value The value to set the Extender channel to
-        
-    */
-    void setElevator(int8_t value);
-    
-    /**
-        
-    Sends a request to the extender to set the Throttle channel to a specific channel between -100 and 100.  
-    
-    @param value The value to set the Throttle channel to
-         
-    */
-    void setThrottle(int8_t value);
-    
-    /**
-         
-    Sends a request to the extender to set the Rudder channel to a specific channel between -100 and 100.  
-    
-    @param value The value to set the Rudder channel to
-         
-    */
-    void setRudder(int8_t value);
-    
-    /**
-         
-    Sends a request to the extender to set the Flight Mode channel to a specific channel between -100 and 100.  
-    
-    @param value The value to set the Flight Mode channel to
-         
-    */
-    void setFlightMode(int8_t value);
-    
-    /**
-         
-    Sends a request to the extender to reply with message containing the current Aileron value.  
-         
-    */
-    void getAileron();
-    
-    /**
-
-    Sends a request to the extender to reply with message containing the current Elevator value. 
-        
-    */
-    void getElevator();
-    
-    /**
-         
-    Sends a request to the extender to reply with message containing the current Throttle value. 
-        
-    */
-    void getThrottle();
-    
-    /**
-        
-    Sends a request to the extender to reply with message containing the current Rudder value. 
-        
-    */
-    void getRudder();
-    
-    /**
-        
-    Sends a request to the extender to reply with message containing the current Flight Mode value. 
-        
-    */
-    void getFlightMode();
-    
-    /**
-         
-    Execute funtion in a thread to not freez the drone.
-         
-    */
-    void aileronCallback(void (*function)(byte));
-    
-    /**
-
-    Execute funtion in a thread to not freez the drone.
-         
-    */
-    void flightModeCallback(void (*function)(byte));
-    
-    /**
-         
-    Execute funtion in a thread to not freez the drone.
-         
-    */
-    void throttleCallback(void (*function)(byte));
-    
-    /**
-         
-    Execute funtion in a thread to not freez the drone.
-         
-    */
-    void rudderCallback(void (*function)(byte));
-    
-    /**
-     
-    Execute funtion in a thread to not freez the drone.
-     
-    */
-    void elevatorCallback(void (*function)(byte));
-    
-    /**
-     
-    Execute funtion in a thread to not freez the drone.
-     
-    */
-    void sendRTEA(uint8_t rudder, uint8_t throttle, uint8_t elevator, uint8_t aileron);
 };
 
 class AvantGPIO
